+++ users.py
@@ -1,4 +1,5 @@
 import datetime as dt
+import json as js
 
 class Users():
     "The class to save user's data..."
@@ -6,29 +7,20 @@ class Users():
     def __init__(self, output_path):
         self.output_path = output_path
 
-    #Saving the user's watchlist...
-    def save_user_list(self, id, the_list):
-        file = open(self.output_path + str(id) + "_watchlist.txt", "w")
-        file.write("exchange;symbol;\n")
-        for symbol in the_list:
-            file.write(symbol[0])
-            file.write(";")
-            file.write(symbol[1])
-            file.write(";\n")
+    #Saving the user's data...
+    def save_user_data(self, id, the_data):
+        file = open(self.output_path + str(id) + ".json", "w")
+        js.dump(the_data, file)
         file.close()
 
-    #To recover the user's watchlist...
-    def load_user_list(self, id):
-        watchlist = None
+    #To recover the user's data...
+    def load_user_data(self, id, chat_data):
         try:
-            file = open(self.output_path + str(id) + "_watchlist.txt", "r").readlines()[1:]
-            watchlist = []
-            for l in file:
-                symbol = l.split(";")
-                watchlist.append((symbol[0], symbol[1]))
+            the_data = js.load(open(self.output_path + str(id) + ".json", "r"))
+            for key in the_data:
+                chat_data[key] = the_data[key]
         except:
             pass
-        return watchlist
 
     #Printing Users()...
     def __str__(self):